--- a/.gitignore
+++ b/.gitignore
@@ -4,3 +4,4 @@
 /deps
 erl_crash.dump
 *.ez
+*.swp