--- a/lib/daily_fantasy.ex
+++ b/lib/daily_fantasy.ex
@@ -101,7 +101,7 @@ defmodule DailyFantasy do
   """
   def filter_by_injury(data) do
     data
-    |> Stream.filter(fn(x) -> x.injury_status in [nil, "P", "Q"] == false end)
+    |> Stream.filter(fn(x) -> x.injury_status in [nil, "", "P", "Q"] end)
   end
 
   @doc """