--- a/lib/daily_fantasy.ex
+++ b/lib/daily_fantasy.ex
@@ -99,7 +99,7 @@ defmodule DailyFantasy do
   Factorial of an integer.
   """
   def factorial(0), do: 1
-  def factorial(n) when n > 0, do: n * fac(n - 1)
+  def factorial(n) when n > 0, do: n * factorial(n - 1)
 
   @doc """
   Quicksort.