--- a/lib/daily_fantasy/lineups.ex
+++ b/lib/daily_fantasy/lineups.ex
@@ -1,0 +1,5 @@
+defmodule DailyFantasy.Lineups do
+  @moduledoc """
+  
+  """
+end