--- a/lib/daily_fantasy/lineups/lineup.ex
+++ b/lib/daily_fantasy/lineups/lineup.ex
@@ -1,0 +1,5 @@
+defmodule DailyFantasy.Lineups.Lineup do
+  @moduledoc """
+  
+  """
+end