Improved efficiency by 33%...

**Before**

    iex(1)>  Benchwarmer.benchmark fn -> DailyFantasy.create_lineups('_data/data.csv') end
    *** #Function<20.52032458/0 in :erl_eval.expr/5> ***
    4.5 sec      1 iterations   4575889.0 μs/op

**After**

    iex(1)>  Benchwarmer.benchmark fn -> DailyFantasy.Lineups.Lineup.FanduelNFL.create_lineups('_data/data.csv') end
    *** #Function<20.52032458/0 in :erl_eval.expr/5> ***
    3.0 sec      1 iterations   3031673.0 μs/op